refactor(Node): rename node attributes
All .rds files for tests had to be regenerated, as the .rds files for the Tree class object could not be loaded. I suspect that readRDS() checks if the class is defined and then verifies its validity. Since I specifically altered the class, it affects the validity of older .rds files. The last commit was not buggy, but this may be because both the previous and new versions of the class were active simultaneously.

I've removed the conversion functions as they can no longer be used.